requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1
black==21.11b1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
django-allauth==0.45.0
django-filter==21.1
django-forms-dynamic==1.0.0
django-heroku==0.3.1
django-htmx==1.11.0
django-mathfilters==1.0.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-widget-tweaks==1.4.12
django==3.2.7
gunicorn==20.1.0
idna==3.2
jinja2==3.0.1
joblib==1.1.0
markupsafe==2.0.1
mypy-extensions==0.4.3
numpy==1.22.4; python_version >= '3.8'
oauthlib==3.1.1
pandas==1.3.4
pathspec==0.9.0
phonenumbers==8.12.46
pillow==8.3.2
platformdirs==2.4.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
pycparser==2.20
pyjwt==2.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
regex==2021.11.10
requests-oauthlib==1.3.0
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.2
setuptools==62.3.2; python_version >= '3.7'
six==1.16.0
sklearn==0.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
whitenoise==5.3.0
